% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzekME.R
\name{analyzekME}
\alias{analyzekME}
\title{analyzes kMEs in comparison and control MEs}
\usage{
analyzekME(datExpr1 = NULL, datExpr2 = NULL, ME_G1 = NULL,
  colorsG1 = NULL, modulesG1 = NULL, species = "Homo.sapiens",
  main = "Gene")
}
\description{
given a control module color set we can test for conservation in a comparison set.  by calculating the signedkME, which is the connectivity of the ModuleEigen values of the first PC1, this will quantify the correlation of the gene and moduleEigenvalues between modules across data sets.
}
