% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzePreservation.R
\name{analyzePreservation}
\alias{analyzePreservation}
\title{analyzes the preservation of modules for a comparison colar assignment}
\usage{
analyzePreservation(geneTree1 = NULL, module1 = NULL, main1 = "",
  geneTree2 = NULL, main2 = "")
}
\arguments{
\item{geneTree1}{the gene tree for the control}

\item{module1}{the module color set for control}

\item{main1}{descriptor}

\item{geneTree2}{the gene tree for the comparison}

\item{main2}{the descriptor}
}
\value{
null
}
\description{
provides a qualitative module preservation between two gene networks
}

