% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDrivers.R
\name{annotateDrivers}
\alias{annotateDrivers}
\title{finds consensus amongst desired traits}
\usage{
annotateDrivers(dbname = NULL, dbPath = ".", consensus = c("Alu", "ERVK",
  "ERV3", "ERVL", "LTR Retrotransposon"), Module.color = "blue",
  species = c("Homo.sapiens", "Mus.musculus"))
}
\value{
a df of the most significant common elements across biotypes per a given module
}
\description{
for traits with similiar correlation patters for a given module, this will find a consensus amongst groups
}
