% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchEffectNormalize.R
\name{batchEffectNormalize}
\alias{batchEffectNormalize}
\title{uses combat to normalize data}
\usage{
batchEffectNormalize(kexp, batch = NULL, design = NULL, read.cutoff = 2,
  byWhich = c("gene_id", "tx_id"), byLevel = c("gene", "repeat"))
}
\value{
batch cpm log2 transformed. and a pdf to show results.
}
\description{
can be called to use edgeR batchEffect removal
}

