% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beePatient.R
\name{beePatient}
\alias{beePatient}
\title{beeswarm patient info plotting cpm or tpm of patient specific, repeat specififc of three interesting clonal processes.}
\usage{
beePatient(kexp, patientID = NULL, repeat_biotype = NULL, selected = 20)
}
\arguments{
\item{kexp}{a kexp}

\item{patientID}{a character to select patient info}

\item{repeat_biotype}{a character either 'element','LTR','SINE', etc}
}
\description{
beeswarm patient info plotting cpm or tpm of patient specific, repeat specififc of three interesting clonal processes.
}

