% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byMad.R
\name{byMad}
\alias{byMad}
\title{selects the top MAD, this was authored by Tim Triche Jr.}
\usage{
byMad(x, k = 500)
}
\arguments{
\item{x}{a matrix}

\item{k}{the number of top MADs to select}
}
\description{
selects the top MAD, this was authored by Tim Triche Jr.
}

