% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGenesAcross.R
\name{compareGenesAcross}
\alias{compareGenesAcross}
\title{this compares genes from across two different experiments}
\usage{
compareGenesAcross(kexp1, kexp2, whichWGCNA = "single",
  species = "Homo.sapiens", read.cutoff = 2, softPower = 10,
  how = "cpm")
}
\description{
comparing genes across different platforms to compare and contrast gene networks across experiments ran on different platforms.
}

