% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRestrictions.R
\name{compareRestrictions}
\alias{compareRestrictions}
\title{compareRestrictions analysis of qusage}
\usage{
compareRestrictions(kexp1, kexp2,
  geneSetPath = "~/Documents/Arkas-Paper-Data/MSigDB/MsigDb_all/",
  compare1 = "Low", compare2 = "pHSC", controls1 = "High",
  controls2 = "LSC", paired1 = FALSE, paired2 = TRUE, read.cutoff = 2,
  MsigDB = NULL, saveToFile = FALSE)
}
\description{
compares the qusage results from two different experiments and finds the correlation of means for each pathway level after standardizing the means.
}

