% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityWiseAnalysis.R
\name{connectivityWiseAnalysis}
\alias{connectivityWiseAnalysis}
\title{tests the connectivity in gene pathways}
\usage{
connectivityWiseAnalysis(lnames, MsigDB = c("c1.all.v5.1.symbols.gmt",
  "c2.all.v5.1.symbols.gmt", "c4.all.v5.1.symbols.gmt",
  "c5.all.v5.1.symbols.gmt", "c6.all.v5.1.symbols.gmt",
  "c7.all.v5.1.symbols.gmt", "h.all.v5.1.symbols.gmt"), pathway = NULL,
  dbname = "wgcnaDBLite.sqlite", qusageDbName = "qusageDbLite",
  gmt.path = "~/Documents/Arkas-Paper-Data/MSigDB/MsigDb_all/",
  moduleColor = "violet")
}
\description{
tests the significance of connectivity of genes in a given pathway resulting in significance tests of connected genes of pathway.
}

