% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumberBiotypeAnalysis.R
\name{copyNumberBiotypeAnalysis}
\alias{copyNumberBiotypeAnalysis}
\title{analyzes the biotypes of the copy number elements}
\usage{
copyNumberBiotypeAnalysis(kexp)
}
\arguments{
\item{kexp}{a kallistoExperiment}
}
\value{
null
}
\description{
using the copy number information, analyzes the biotypes of each copy number and ranks them
}

