% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumberNormalize.R
\name{copyNumberNormalize}
\alias{copyNumberNormalize}
\title{normalizes the copy number for repeat biotypes family expression}
\usage{
copyNumberNormalize(kexp, bundleID = "tx_biotype", read.cutoff = 1,
  discardjoined = TRUE)
}
\value{
a matrix of copy normalized repeat counts
}
\description{
averages the family biotype expression and normalizes with the average copy number
}

