% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqnDE.R
\name{cqnDE}
\alias{cqnDE}
\title{CQN and edgeR analysis}
\usage{
cqnDE(kexp, contrastMatrix = NULL, design = NULL, patientID = NULL,
  comparison, control, group3)
}
\arguments{
\item{kexp}{a kallistoExperiment preferably only repeats, also the repeat kexp must not be TMM Normalized}

\item{contrastMatrix}{a matrix with contrasts, this is specific and this function is created without fail safes}

\item{design}{a matrix to compare, 2 group treatment factor}

\item{comparison}{the column group either pHSC , LSC ,or Blast}

\item{control}{not comparison}

\item{group3}{not comparison and not control}
}
\value{
returns
}
\description{
this runs CQN and edgeR on a kallistoExperiment with dynamic contrasts to select  CQN normalization across clonal stages.  This must have replicates so we are comparing samples and not specific patients.
}

