% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqnRepeatAnalysis.R
\name{cqnRepeatAnalysis}
\alias{cqnRepeatAnalysis}
\title{cqnDE analysis on repeats, major differences in performance when CQN is called on transcripts and/or called on repeat elements. ALUs have higher GC conetnet and may have bias}
\usage{
cqnRepeatAnalysis(fullKexp, cqnRepeat = NULL, inversion = TRUE,
  comparison = comparison, control = control)
}
\arguments{
\item{fullKexp}{a full kallistoExperiment, by Stage is ok}

\item{cqnRepeat}{a data matrix of cqnDE result initially from a full expression log2 scale of all transcripts, a full normalization}

\item{inversion}{boolean , if true then the log2 inverse is performed}
}
\value{
returns some images on comparing repeat CQN method
}
\description{
this methods is downstream of cqnDE where it looks at repeat subset expression from the cqnDE entire transcript set.  we are not sure if the cqn method has a bias with regard to repeats, and we are not sure if repeats with sequence repetitive nature can bias the cqn method when **only** normalizing with respect to repeats.   so cqnDE is called with the entire transcript set, and the repeats are merely subset out; contrast to calling cqn on repeats only and analyzing the result.
}

