% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendriticWiseAnalysis.R
\name{dendriticWiseAnalysis}
\alias{dendriticWiseAnalysis}
\title{analyzing the dendritic tet2 KO experiment requires simulating replicates}
\usage{
dendriticWiseAnalysis(kexp, bundleID = "tx_id", read.cutoff = 1,
  whichModel = c("norm", "pois", "gamma", "exp", "unif"), n.sims = 50000,
  minTPM = 0.3, newCols = 3)
}
\arguments{
\item{kexp}{a kallisto experiment repeat stage}

\item{bundleID}{tx_id or gene_id}

\item{read.cutoff}{floor threshold}

\item{whichModel}{norm pois gamma exp or unif, these are options to test the fit model against}

\item{n.sims}{5e4 numeric to simulate trials total}

\item{minTPM}{for exp model TPMs are used and this is the floor TPM cut}

\item{newCols}{integer, new columns to simulate: breathe in the newness of life.}
}
\value{
a matrix/df of simulated columns dim n X (2*newCols+2)
}
\description{
the tet2 KO has no replicates so this method will simulate samples using a uniform distribution to create multiple replicates with some technical digital variation.  a DE call will be used but the results are from simulated data.
}

