% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawBatchHeatmap.R
\name{drawBatchHeatmap}
\alias{drawBatchHeatmap}
\title{Draws a heatmap with annotations for batch data}
\usage{
drawBatchHeatmap(kexp, tags = NULL, annotations = TRUE, batchData = NULL,
  cutoff = 0.05, byWhat = c("counts", "tpm"))
}
\arguments{
\item{kexp}{a kallistoExperiment repeat stage level}

\item{tags}{character of top tags from DE list}

\item{annotations}{boolean  to include annotation side bar, or mutation type}

\item{batchData}{this is patient specific batch corrections}
}
\description{
draws a heatmap with annotations useful for batch analysis normalization. plots the batch normalized data values, and not the raw kallisto data.
}

