% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDensityPlots.R
\name{drawDensityPlots}
\alias{drawDensityPlots}
\title{Density plots and custom imaging}
\usage{
drawDensityPlots(kexp, comparison = "pHSC", control = "LSC",
  comparison2 = "Blast", numberComparisons = 1, read.cutoff = 3,
  adjustBy = "BH", wilcox.Alternative = c("less", "greater"),
  grob.comparison1 = c("Md(High Grade)<Md(Low Grade) p.value: "),
  grob.comparison2 = c("Md(High Grade)<Md(Low Grade p.value: "),
  testMedians = FALSE)
}
\arguments{
\item{kexp}{this is the non-stage level kexp for 3 stages}

\item{comparison}{the comparison group}

\item{control}{the reference group}

\item{comparison2}{if more than one comparison}

\item{numberComparisons}{integer, the number of comparisons}

\item{read.cutoff}{integer, to filter for repeatWise Analysis}

\item{adjustBy}{either none,BH,BY, or holm for lmFit call}

\item{wilcox.Alternative}{either less, which is testing the alternative that the comparison is less than the control, or greater which tests the alternative hypothesis that the comparison is greater than control.}

\item{grob.comparison1}{this is the text printed to plot that verbose the alternative}

\item{grob.comparison2}{this is the second printed explanation of the alternative hypothesis printed on plot testing medians}

\item{testMedians}{boolean , whether to include a median test of logFC from the repeat linear modeling.}
}

