% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDF.R
\name{drawDF}
\alias{drawDF}
\title{draws a heatmap downstream of PoissonSeq}
\usage{
drawDF(kexp, res = res, p.value = 0.05, cutoff = 1, fromList = FALSE)
}
\arguments{
\item{kexp}{a kallisto experiment usually a repeat stage level}

\item{res}{poissonSeq fit}

\item{p.value}{numeric floor}
}
\value{
images and stuffs
}
\description{
draws a heatmap from the poissonseq DE list
}

