% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.r
\name{dt_quantilebins_generic}
\alias{dt_quantilebins_generic}
\title{Generic quantile binning}
\usage{
dt_quantilebins_generic(
  DT,
  numerical.var,
  weight.var,
  by.vars = NULL,
  outvarnames,
  wt.quantile.fun,
  ...
)
}
\arguments{
\item{DT}{Data table to be summarized}

\item{numerical.var}{\link{str(1)} which variable will be quantilized}

\item{weight.var}{\link{num} weights with which quantiles are constructed}

\item{by.vars}{\link{str(n)} vector of variables by which}

\item{outvarnames}{not implemented yet: control return names}

\item{wt.quantile.fun}{\link{function} f(x,w) where x and w have equal length}

\item{...}{extra arguments to \code{cut}}
}
\value{

}
\description{
Generic function for quantile binning
}
\details{
Creates quantiles of variable \code{numerical.var} (supplied as a single string)
using weights given by \code{weight.var}.
}
\author{
Gustavo
}
