% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.r
\name{dt_quantilebins_weighted}
\alias{dt_quantilebins_weighted}
\title{Weighted quantile bins from data.table data}
\usage{
dt_quantilebins_weighted(
  DT,
  numerical.var,
  wt.var,
  by.vars = NULL,
  ntile = 5,
  outvarnames = c("bin", "bin_n"),
  bounds = c(-Inf, Inf),
  ...
)
}
\arguments{
\item{DT}{\link{data.table} data}

\item{numerical.var}{\link{str(1)} variable to be quantilized}

\item{wt.var}{\link{str(1)} variable to be used as weight}

\item{by.vars}{\link{str(n)} variable to subset by when computing quantiles}

\item{ntile}{\link{int} number of quantiles}

\item{outvarnames}{to be implemented}

\item{bounds}{\link{num(2)} natural bounds for \code{numerical.var} (for pretty printing)}

\item{...}{extra arguments to the \code{cut} function}
}
\value{

}
\description{
Compute weighted quantiles of a data.table \code{DT}
}
\details{
Apply \code{dt_quantilebins_generic} with Hmisc's \code{wtd.quantile} to find
weighted quantiles. The number of percentiles is governed by \code{ntile}: with
\code{ntile=5}, the function computes quintiles, with \code{ntile=4} quartiles,and so
on.

The \code{...} are passed to the \code{cut} function. Note that \code{ordered_result=TRUE}
necessarily (i.e., it's not an option).
}
\author{
Gustavo
}
