% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensmblToHUGO.R
\name{ensmblToHUGO}
\alias{ensmblToHUGO}
\title{map ensembl id to hugo via biomaRt}
\usage{
ensmblToHUGO(kexp, commonNomen = c("human", "mouse"), byWhich = c("gene",
  "repeat"))
}
\arguments{
\item{kexp}{a grimes kexp at the gene or repeat level}

\item{commonNomen}{character moues or human}
}
\value{
a list with the meta data added
}
\description{
sometimes you just want gene names from a TPM list.  this uses a collapseTPM call to map to HUGO to allow for exploration
}
