% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInterestingModules.R
\name{findInterestingModules}
\alias{findInterestingModules}
\title{read key for inputing interesting modules}
\usage{
findInterestingModules(n.modules = 7)
}
\description{
read key for inputing interesting modules
}
