% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRepeats.R
\name{findRepeats}
\alias{findRepeats}
\title{splits a kexp into repeat level}
\usage{
findRepeats(kexp)
}
\arguments{
\item{kexp}{a kallistoExperiment class object assumes pre-annotated}
}
\value{
returns a repeat level kallisto Experiment
}
\description{
splits a kexp into repeat level
}
