% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherTestCorrelations.R
\name{fisherTestCorrelations}
\alias{fisherTestCorrelations}
\title{tests the correlation and enrichment association from the correlation map}
\usage{
fisherTestCorrelations(lnames = NULL, rnames = NULL, read.cutoff = 1,
  recalc = FALSE, how = how, pathwaysToPick = c("Immune", "Inflammation"),
  pathPairing = c(1, 1), dbname = NULL, qdbname = NULL, rdbname = NULL,
  p.value = 0.1, minRank = 0)
}
\arguments{
\item{lnames}{this is the results from wgcna.R}

\item{read.cutoff}{integer for min cutoff}

\item{recalc}{boolean if truen then will recalculate the bicor and pvalues}

\item{how}{tpm or cpm, this is used to print the heatmap, tpm is better}

\item{pathwaysToPick}{this will query the qusageDbLite database for the key pathway character name signature}

\item{pathPairing}{vector if pathways are related then they should have matching pairing integers for instance cell death and apoptosis pathways could be paired.}

\item{dbname}{the gene module wgcnaDbLite sqlite database}

\item{qdbname}{the qusageDbLite sqlite database name}

\item{rdbname}{the repeat module wrcnaDbLite sqlite database}

\item{p.value}{this is the p.value threshold for selecting the inferred function of a given module.}

\item{minRank}{In a given module it can filter based on the competitive rank of where the pathwaysToPick are ranked in a given module.}
}
\value{
images of eigengenes
}
\description{
after calling CorrelationHeatmap we can test the correlations association to the enrichment activation. This uses a fisher test to test each TE type individually association to the TE type's enrichment activation directions.  finally we aggregate all the TE types count data and perform a global chi-sq test for an overall association.
}
