% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTxBiotypes.R
\name{fitTxBiotypes}
\alias{fitTxBiotypes}
\title{plots the stage DE biotype}
\usage{
fitTxBiotypes(kexp, design = NULL)
}
\arguments{
\item{kexp}{a kallisot Experiemtn at the stage repeat level}
}
\value{
a bunch of useful plots
}
\description{
plots the stage of each DE
}
