% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitWithoutReplicates.R
\name{fitWithoutReplicates}
\alias{fitWithoutReplicates}
\title{this analysis dendritic cells from Zhang et. al}
\usage{
fitWithoutReplicates(kexp, bundleID = "tx_id", read.cutoff = 1,
  byWhat = c("genes", "repeats"), how = c("counts", "tpm"), ...)
}
\arguments{
\item{kexp}{a kexp of dendritic 2 group}

\item{bundleID}{collapse range column}

\item{read.cutoff}{floor expression}
}
\value{
a DGE list
}
\description{
dendritic tet2 cells in a 2 group study. this analyzes 2 groups 1 comparison and 1 control. this performs an MDS plot of the dendritic 4h/0h
}

