% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixRepeatRanges.R
\name{fixRepeatRanges}
\alias{fixRepeatRanges}
\title{adds repeat annotations to gene level analysis}
\usage{
fixRepeatRanges(kexp)
}
\value{
a kexp
}
\description{
allows for gene level analsysis competitive against genes and not only transcripts
}

