% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_the_top.R
\name{from_the_top}
\alias{from_the_top}
\title{A add in that runs a clears the working environment and packages and runs from the top}
\usage{
from_the_top()
}
\value{
The environment a script run top to bottom
}
\description{
A add in that runs a clears the working environment and packages and runs from the top
}
