% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_the_top.R
\name{from_the_top_bkgd}
\alias{from_the_top_bkgd}
\title{A add in that runs a script in the background but doesn't return output}
\usage{
from_the_top_bkgd()
}
\value{
Everything is run in the background and nothing is output to the global environment
}
\description{
A add in that runs a script in the background but doesn't return output
}
