% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_the_top.R
\name{from_the_top_bkgd_output}
\alias{from_the_top_bkgd_output}
\title{A add in that runs a script in the background and returns output to global environment}
\usage{
from_the_top_bkgd_output()
}
\value{
Everything is run in the background and output is returned to global environment
}
\description{
A add in that runs a script in the background and returns output to global environment
}
