% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcNorm.R
\name{gcNorm}
\alias{gcNorm}
\title{GC Normalize using conditional quantile normalization}
\usage{
gcNorm(kexp)
}
\arguments{
\item{kexp}{kallistoExperiment object}
}
\value{
a kallistoExperiment object with GC normalized assays added
}
\description{
this uses conditional quantile normalization an EDASeq to normalize and reduce GC bias
}

