% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotation.R
\name{geneAnnotation}
\alias{geneAnnotation}
\title{creates a geneInfo0 data frame which has gene symbol and module color}
\usage{
geneAnnotation(datExpr, datTraits, species = c("Homo.sapiens",
  "Mus.musculus"))
}
\value{
geneInof0 data frame
}
\description{
used mainly for WGCNA which returns a data frame of geneSym locus link ID module colr and correlations
}

