% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneListToGMT.R
\name{geneListToGMT}
\alias{geneListToGMT}
\title{convert list of genes to GMT format for qusage calls}
\usage{
geneListToGMT(geneList = NULL, fileName = "test.gmt",
  citation = "www.na.com")
}
\arguments{
\item{geneList}{a list of genes with names(list)}

\item{fileName}{output.gmt}

\item{citation}{the source of gene list}
}
\description{
given a gene list write to a .gmt file
}
