% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneTest.R
\name{geneTest}
\alias{geneTest}
\title{this creates an expression time plot of match patients}
\usage{
geneTest(kexp, geneName = "EVI", how = c("cpm", "tpm"), saveOut = FALSE,
  numberPairs = c(1, 2, 3), read.cutoff = 1)
}
\arguments{
\item{kexp}{this is a kexp for 2 gropu non-paired patient wise, with 2 groups}
}
\value{
images and a pdf/jpeg
}
\description{
for non-paired group factor ANOVA testing, this plots the expression across two groups units of TPM
}

