% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeeswarm.R
\name{getBeeswarm}
\alias{getBeeswarm}
\title{runs a beeswarm plot}
\usage{
getBeeswarm(name, dataMat, what = c("tpm", "count"), phenoData, patientID,
  repeat_biotype, how)
}
\description{
runs a beeswarm plot
}

