% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_name_split.R
\name{group_name_split}
\alias{group_name_split}
\title{Splits a data frame into a named list given grouping variables}
\usage{
group_name_split(.tbl, ..., .keep = TRUE, .sep = "_")
}
\arguments{
\item{.tbl}{A tbl.}

\item{...}{If .tbl is an ungrouped data frame, a grouping specification, forwarded to dplyr::group_split().}

\item{.keep}{Should the grouping columns be kept?}

\item{.sep}{Separator to use between grouping specification.}
}
\value{
A named list of data frames
}
\description{
Splits a data frame into a named list given grouping variables
}
\examples{
data <- data.frame(first = c('jonathan','porter','sam','davis','quinn'),
                   last = c('duncan','christensen','morgan','dayley','durrant'),
                   x = rnorm(100), y = rnorm(100,2), z = rnorm(100,10,5))

group_name_split(data,first)

}
