% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapByMad.R
\name{heatmapByMad}
\alias{heatmapByMad}
\title{draws a heatmap by MAD}
\usage{
heatmapByMad(data, kexp, topAnnoFactors = NULL, selectK = NULL,
  byWhat = byWhat)
}
\arguments{
\item{data}{a matrix of data values}

\item{kexp}{a kallistoExperiment by any stage, or could be repeats, general use}

\item{topAnnoFactors}{a factor of top annotation bars}
}
\value{
returns images
}
\description{
a general function for heatmapping data by MAD
}

