% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kexp2Group.R
\name{kexp2Group}
\alias{kexp2Group}
\title{splits kexp by group  pHSC , LSC, Blast}
\usage{
kexp2Group(kexp, comparison = NULL, control = NULL)
}
\arguments{
\item{kexp}{a kallisto experiment}

\item{comparison}{pHSC LSC etc}

\item{control}{not comparions}
}
\description{
splits kexp by group  pHSC , LSC, Blast
}

