% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kexp3GroupTo2.R
\name{kexp3GroupTo2}
\alias{kexp3GroupTo2}
\title{splits kexp by group  pHSC , LSC, Blast}
\usage{
kexp3GroupTo2(kexp, comparison = NULL, newComparison = NULL)
}
\arguments{
\item{kexp}{a kallisto experiment}

\item{comparison}{pHSC LSC etc}

\item{control}{not comparions}
}
\description{
splits kexp by group  pHSC , LSC, Blast
}

