% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kexpAnalysis.R
\name{kexpAnalysis}
\alias{kexpAnalysis}
\title{analyzes kexp data}
\usage{
kexpAnalysis(kexp, byWhich = c("gene", "repeat"), read.cutoff = 1,
  adjustBy = c("BH", "none"))
}
\arguments{
\item{kexp}{mpn vs Mx Cre lines}
}
\description{
analyzes kexp repeat and gene data but requires a design matrix.  uses RUV and a limma/voom model linear
}

