% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kexpByPatient.R
\name{kexpByPatient}
\alias{kexpByPatient}
\title{splits kexp by patient ID}
\usage{
kexpByPatient(kexp, patientID = NULL)
}
\arguments{
\item{kexp}{a kallisto experiment}

\item{patientID}{a patient id to split columns}
}
\description{
splits kexp by patient ID
}

