% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kexpByStage.R
\name{kexpByStage}
\alias{kexpByStage}
\title{splits a kallisto experiment by stages}
\usage{
kexpByStage(kexp, stage1 = "pHSC", stage2 = "LSC", stage3 = "Blast")
}
\arguments{
\item{kexp}{a kallisto Experiment}

\item{stage1}{first stage}

\item{stage2}{second stage}

\item{stage3}{third stage}
}
\description{
splits a kallisto experiment by stages
}

