% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kexpByTrio.R
\name{kexpByTrio}
\alias{kexpByTrio}
\title{this will subset a kexp by patient trio data}
\usage{
kexpByTrio(kexp, stage1 = "pHSC", stage2 = "LSC", stage3 = "Blast")
}
\description{
given a trio time series we subset a kexp so that each patient trio exists across all time points
}
