% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kexpByType.R
\name{kexpByType}
\alias{kexpByType}
\title{splits a kallisto experiment by gene biotype}
\usage{
kexpByType(kexp, biotype = NULL)
}
\arguments{
\item{kexp}{a kallisto Experiment}

\item{biotype}{character a selector either 'element','SINE','LINE','LTR', etc}
}
\description{
splits a kallisto experiment by gene biotype
}
