% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeTable.R
\name{kmeTable}
\alias{kmeTable}
\title{calculates the connectivity of the Module eigenvalues across genes}
\usage{
kmeTable(datExpr1 = NULL, ME_G1 = NULL, colorsG1 = NULL,
  modulesG1 = NULL, species = "Homo.sapiens", main = "Gene")
}
\arguments{
\item{datExpr1}{expression data}

\item{ME_G1}{module eigenvalues for the expression data}

\item{colorsG1}{the color assignment}

\item{modulesG1}{the partition of the expression data}

\item{species}{Homo.sapiens}

\item{main}{gene or Repeat}
}
\description{
the module eigenvalues are defined as the variability of a given module, and the kME is the connectivity of genes to the Module eigenvalues.  if genes are highly connected to the moduleseigen values then they will behave connectedly to the gene corresponding to the ME.
}
