% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laneNorm.R
\name{laneNorm}
\alias{laneNorm}
\title{performs EDASeq  lane normalization}
\usage{
laneNorm(kexp)
}
\arguments{
\item{kexp}{KallistoExperiment}
}
\description{
Peforms within lane and between lane normalization which has been shown to have a lower Type 1 error performed in yeast by Risso et. al.
}

