% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levineAnalysis.R
\name{levineAnalysis}
\alias{levineAnalysis}
\title{analyzes levine data}
\usage{
levineAnalysis(kexp)
}
\value{
a kexp analysis
}
\description{
analyzes the levine kexp by GMP Mx-Cre+ (interferon activated) (3) vs Mx-Cre (neg) Cntrl.
}

