% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFunctionDataFrame.R
\name{moduleFunctionDataFrame}
\alias{moduleFunctionDataFrame}
\title{given the module of interest printed to a csv, we can read that data, and analyze the pathway behavior of specific pathways}
\usage{
moduleFunctionDataFrame(db = NULL, qdb = NULL, functionKeyWords = NULL)
}
\arguments{
\item{db}{the wgcnaDbName}

\item{qdb}{the qusageDbLite name}

\item{functionKeyWords}{this is a single character that will be grep'd from the list of pathway modules}
}
\value{
a data frame with the queried keyword in every module and pathway information
}
\description{
reads the qusage module pathway activity and analyzes specific behvaviors within that function
}

