% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_heatmap.R
\name{moduleHeatmap}
\alias{moduleHeatmap}
\title{used in wgcna_significance}
\usage{
moduleHeatmap(datExpr, colorh1, intModules)
}
\arguments{
\item{datExpr}{a data expression}

\item{colorh1}{a color vector labels}

\item{intModules}{a vector of length 3}
}
\value{
cool heatmap images of modules over samples
}
\description{
the wgcna_significane has user input to investigate interesting modules. these intModules are saved during this script.  wgcna_heatmap will produce the datExpr values across all samples
}

