% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleHubGenes.R
\name{ModuleHubGenes}
\alias{ModuleHubGenes}
\title{returns the central hub genes for analysis within a kexp for a selected module color. topHubGenes finds hubs across all modules.}
\usage{
ModuleHubGenes(kexp, lnames = NULL, module.color = "1",
  connectivity.cutoff = 0.6)
}
\arguments{
\item{kexp}{kexp that is annotated}

\item{lnames}{wgcna.R output}

\item{module.color}{an integer or character specifying the color, "blue", or the module number.  if ME column names are MM.1, then use the integers.  otherwise, the ME labels are characters, requiring "blue" input.}
}
\value{
a list of genes and their connectivity
}
\description{
performs a connectivity analysis and returns the top connected genes within an experiment.
}
