% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePCA.R
\name{modulePCA}
\alias{modulePCA}
\title{this will plot a cluster of eigen values and pca of module eigenvalues}
\usage{
modulePCA(lnames)
}
\arguments{
\item{lnames}{this is the wgcna.cpm.RData returned from wgcna.R}
}
\description{
by clustering moduleEigenvalues one can examine how close other modules are to one another.  if to module eigenvalues are branched in the same tree, then their functions pathway are likely to be similiar.  Further modules with similiar eigenvalues can even be merged.
}

