% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleWiseAnalysis.R
\name{moduleWiseAnalysis}
\alias{moduleWiseAnalysis}
\title{this analyzes correlations of repeat and gene modules of fixed interest}
\usage{
moduleWiseAnalysis(kexp, lnames, rnames,
  wgcnaDbName = "wgcnaDbLite.cpm.sqlite",
  wrcnaDbName = "wrcnaDbLite.cpm.sqlite",
  qusageDbName = "qusageDbLite.cpm.sqlite", geneModules = c("darkturquoise",
  "red", "grey60", "pink", "darkred", "royalblue", "green", "magenta", "brown"),
  repeatModules = c("blue", "brown", "grey", "turquoise"), read.cutoff = 2,
  how = c("cpm", "tpm"))
}
\description{
after running wgcna and wrcna we are interested in finding associaitons between gene and repeat modules.  given modules of interest we correlate and test for comparability between each module.  we also analyze the associations of connectivity.
}

