% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeByCopyNumber.R
\name{normalizeByCopyNumber}
\alias{normalizeByCopyNumber}
\title{each repeat has lots of copy number by nature, so we normalize by copy number to have expression count per copy}
\usage{
normalizeByCopyNumber(kexp, species = c("Homo.sapiens", "Mus.musculus"))
}
\arguments{
\item{kexp}{a kallistoExperiment}
}
\value{
a matrix or data frame of copy normalized counts and tpm
}
\description{
divides the repeat count by copy number
}

