% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omitRelapsed.R
\name{omitRelapsed}
\alias{omitRelapsed}
\title{omits SU583 from kexp}
\usage{
omitRelapsed(kexp, relapsedID = c("SU351", "SU583"))
}
\arguments{
\item{kexp}{a kexp at any level with SU583 as column}

\item{relapsedID}{character SU351 and SU583 are reported relapsed, removing}
}
\value{
a kexp
}
\description{
SU583 donor was not taken the same as the others but was taken post relapse, this sample should be omitted.
}

