% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omitSU583.R
\name{omitSU583}
\alias{omitSU583}
\title{omits SU583 from kexp}
\usage{
omitSU583(kexp)
}
\arguments{
\item{kexp}{a kexp at any level with SU583 as column}
}
\value{
a kexp
}
\description{
SU583 donor was not taken the same as the others but was taken post relapse, this sample should be omitted.
}

