% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderDEMonotonicity.R
\name{orderDEMonotonicity}
\alias{orderDEMonotonicity}
\title{finds the monotonic nominal ordering group for DE lists}
\usage{
orderDEMonotonicity(mt, globalMax = stageId1, globalMin = stageId2,
  globalMid = stageId3)
}
\arguments{
\item{mt}{a matrix of DE lists from limma}

\item{globalMax}{the selected global max column number}

\item{globalMin}{the selected global min column number}

\item{globalMid}{the column number of desired mid}
}
\value{
a matrix group that fits into this nominal group ordering
}
\description{
for a given stage the ordering is selected for DE
}

