% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderIncreasing.R
\name{orderIncreasing}
\alias{orderIncreasing}
\title{orders the increasing  nominal group}
\usage{
orderIncreasing(mt, globalMax = stageId1, globalMin = stageId2,
  globalMid = stageId3, increaseStage = stageId2)
}
\arguments{
\item{mt}{a matrix of collapse bundles kexp}

\item{globalMin}{the selected global min column number}

\item{globalMid}{the column number of desired mid}

\item{increaseStage}{is either the first or second stage with constant othewise}
}
\value{
a matrix group that fits into this nominal group ordering
}
\description{
orders the change in clonal stage as increasing between specific clonal stage 2 or 3, where pre-leukemic stage is low
}

