% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderMonotonicity.R
\name{orderMonotonicity}
\alias{orderMonotonicity}
\title{finds the non monotonic nominal ordering group}
\usage{
orderMonotonicity(mt, globalMax = stageId1, globalMin = stageId2,
  globalMid = stageId3)
}
\arguments{
\item{mt}{a matrix of collapse bundles of repeats}

\item{globalMax}{the selected global max column number}

\item{globalMin}{the selected global min column number}

\item{globalMid}{the column number of desired mid}
}
\value{
a matrix group that fits into this nominal group ordering
}
\description{
for a given stage the ordering is selected
}

