% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientBatchEffects.R
\name{patientBatchEffects}
\alias{patientBatchEffects}
\title{this analyzes the batch effects from specific patients, where each patient ID is considered a batch}
\usage{
patientBatchEffects(kexp, byWhat = c("counts", "tpm"), stage1 = "pHSC",
  stage2 = "LSC", stage3 = "Blast", p.value = 0.05)
}
\arguments{
\item{kexp}{a full kexp of every sample, not just a stage kexp}

\item{byWhat}{how to collapse}
}
\value{
returns images and a patient batch corrected data set
}
\description{
performs a batch correction where each patient is a batch so to remove out individual patient effects to make each donor uniform.  it is critical to put the entire 67 samples kexp into the input.
}

