% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientPlot.R
\name{patientPlot}
\alias{patientPlot}
\title{plots a heatmap of patient specific information using a heatmap and beeswarm for the scatter.  each plot shows fixed patient and fixed repeat biotype}
\usage{
patientPlot(kexp, patientID = NULL, repeatType = NULL)
}
\arguments{
\item{kexp}{a kallisto Experiment}

\item{patientID}{a patient id to split a kexp by}

\item{repeatType}{a repeat biotype of interest, 'LTR', 'SINE', 'element', etc}
}
\value{
returns NULL but plots cool stuff
}
\description{
plots a heatmap of patient specific information using a heatmap and beeswarm for the scatter.  each plot shows fixed patient and fixed repeat biotype
}

