% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientPlot_cpm_library_norm.R
\name{patientPlot_cpm_library_norm}
\alias{patientPlot_cpm_library_norm}
\title{plots a heatmap of patient specific information using a heatmap and beeswarm for the scatter.  each plot shows fixed patient and fixed repeat biotype of cpm TMM normalized kallistoExperiments}
\usage{
patientPlot_cpm_library_norm(kexp, patientID = NULL, repeatType = NULL,
  normType = c("TMM", "CQN", "none"))
}
\arguments{
\item{kexp}{a kallisto Experiment}

\item{patientID}{a patient id to split a kexp by}

\item{repeatType}{a repeat biotype of interest, 'LTR', 'SINE', 'element', etc}

\item{normType}{either TMM CQN or none}
}
\value{
returns NULL but plots cool stuff
}
\description{
plots a heatmap of patient specific information using a heatmap and beeswarm for the scatter.  each plot shows fixed patient and fixed repeat biotype of cpm TMM normalized kallistoExperiments
}

