% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientTrioPlot.R
\name{patientTrioPlot}
\alias{patientTrioPlot}
\title{this creates an expression time plot of match patients}
\usage{
patientTrioPlot(kexp, stage1 = "pHSC", stage2 = "LSC", stage3 = "Blast",
  printWhat = "jpeg", printOut = FALSE)
}
\arguments{
\item{kexp}{this is a kexp for three stages patient wise, with 3 time-stages}

\item{stage1}{character pHSC}

\item{stage2}{character LSC}

\item{stage3}{character Blast}
}
\value{
images and a pdf/jpeg
}
\description{
for paired time series, this plots the expression across time points in units of TPM. This plots each repeat biotype and every patient trio across stages.
}
