% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientWiseAnalysis.R
\name{patientWiseAnalysis}
\alias{patientWiseAnalysis}
\title{patient Differential Expression analysis}
\usage{
patientWiseAnalysis(kexp, whichDE = c("cqn", "poissonSeq", "limma"),
  comparison = "LSC", control = "pHSC", whichDelta = c("delta1",
  "delta2"), patientID = "SU353", orderDE = TRUE)
}
\arguments{
\item{kexp}{for DE poissonSeq kexp should repeat stage, for cqn kexp should be stage transcrpit}

\item{whichDE}{the de caller cqn, poissonSeq limma , limma should call stageWiseaNALYSIS}

\item{comparison}{character the comparison used for 2 group cmpr/control}

\item{control}{the character in the denominator of 2 group DE call}

\item{whichDelta}{characther  either LSC/pHSC or Blast/LSC}

\item{patientID}{character  ID to subset}

\item{orderDE}{boolean ,  if true the TPMs of the topNamed repeat tags are plotted,  if FALSE the TPMs are all plotted for global view}
}
\value{
a list of all the monotonic ordinal groups
}
\description{
patientPlot* is exploratory methods for expression patterns, this method is downstream of a selected DE caller. the inputs are a repeat stage level kexp, and the DE object from either cqnDE, poissonFit, edgeR, or limma calls.  The topSelected repeats are used to plot the overall heatmap, and then heatmap each tx_biotype separately.   the topSelected are then categorized based on ordinal groups, and the group frequencies are plotted.  pvalue histograms are included.  delta plots are included (logFC).
}
