% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peggyTet2Analysis.R
\name{peggyTet2Analysis}
\alias{peggyTet2Analysis}
\title{this analyzes tet2 and dnmt3a dKO HSC}
\usage{
peggyTet2Analysis(kexp)
}
\arguments{
\item{kexp}{this is the full kexp from peggy's HSC, we will split into each group and run a linear model regression on the data for repeats}
}
\description{
title
}

