% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBatchFrequency.R
\name{plotBatchFrequency}
\alias{plotBatchFrequency}
\title{plot frequency of biotypes post batch correction}
\usage{
plotBatchFrequency(dataMat, topNames = NULL, stage1 = "pHSC",
  stage2 = "LSC", stage3 = "Blast", topDE = NULL,
  whichDelta = c("delta1", "delta2"), p.cutoff = 0.05, isAdjusted = FALSE,
  kexp = NULL)
}
\arguments{
\item{dataMat}{a batch correction psva data matrix}

\item{topNames}{names of the top DE list of PVal or adj}

\item{stage1}{first stage}

\item{stage2}{second stage}

\item{stage3}{third stage}

\item{topDE}{differential expr list}

\item{whichDelta}{significant DE is a change between states, so which state is being measured?}

\item{p.cutoff}{numeric p value threshold}

\item{isAdjusted}{boolean if true than the adj.P.Values are used and labels with show accordingly}

\item{kexp}{this is a repeat stage kexp, used only for grabbing annotations}
}
\value{
cool images
}
\description{
plots the histogram and bar charts of biotypes after batch correcting
}

