% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFrequency.R
\name{plotFrequency}
\alias{plotFrequency}
\title{plot frequency of biotypes}
\usage{
plotFrequency(kexp, topNames = NULL, stage1 = "pHSC", stage2 = "LSC",
  stage3 = "Blast", topDE = NULL, whichDelta = c("delta1", "delta2"),
  p.cutoff = 0.05, isAdjusted = FALSE)
}
\arguments{
\item{kexp}{a kallistoExperiment}

\item{topNames}{names of the top DE list of PVal or adj}

\item{stage1}{first stage}

\item{stage2}{second stage}

\item{stage3}{third stage}

\item{topDE}{differential expr list}

\item{whichDelta}{significant DE is a change between states, so which state is being measured?}

\item{p.cutoff}{numeric p value threshold}

\item{isAdjusted}{boolean if true than the adj.P.Values are used and labels with show accordingly}
}
\value{
cool images
}
\description{
plots the histogram and bar charts of biotypes
}

