% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMonotonicity.R
\name{plotMonotonicity}
\alias{plotMonotonicity}
\title{plots the monotonic order}
\usage{
plotMonotonicity(kexp, nominalGroup = NULL, stageId1 = stageId1,
  stageId2 = stageId2, stageId3 = stageId3, mt, read.cutoff = read.cutoff,
  stage3 = stage3, Max = globalMax, Mid = globalMid, Min = globalMin)
}
\arguments{
\item{kexp}{a kallisto Experiment patient repeat level}

\item{nominalGroup}{a monotonic order}

\item{stageId1}{the first clone stage}

\item{stageId2}{the second clone stage}

\item{stageId3}{the third stage}

\item{mt}{a collapse bundles matrix}

\item{read.cutoff}{the cutoff for CPM min floor}

\item{Max}{a column of a kexp}

\item{Mid}{a column of a kexp}

\item{Min}{a column of a kexp}
}
\value{
some plots
}
\description{
plots the monotonic ordering
}

