% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonFit.R
\name{poissonFit}
\alias{poissonFit}
\title{fits a poisson model to determine goodness of fit}
\usage{
poissonFit(kexp, bundleID = "tx_id", cutoff = 1, comparison = "LSC",
  control = "pHSC", p.value = 0.05, whichDelta = c("delta1", "delta2"))
}
\arguments{
\item{kexp}{kallisto experiment at repeat stage level}
}
\value{
data and some images fit to poisson log linear model
}
\description{
determines if the fit is poisson distributed and fits a poisson regression model
}

