% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerThresholdAnalysis.R
\name{powerThresholdAnalysis}
\alias{powerThresholdAnalysis}
\title{this will analysze the power threshold beta}
\usage{
powerThresholdAnalysis(lnames = NULL, datExpr = NULL, how = "gene")
}
\arguments{
\item{lnames}{this is the list from wgcna.R}

\item{datExpr}{this is the log2 XR rows as samples columns as genes/repeats}

\item{how}{a character param for pdf print}
}
\description{
although the network is robust to the choice of beta, this method will call the pickSoftThreshold values and print to a pdf for future use.
}
