% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyPreservation.R
\name{quantifyPreservation}
\alias{quantifyPreservation}
\title{quantifies preservation of modules}
\usage{
quantifyPreservation(datExpr1 = NULL, datExpr2 = NULL, modules = NULL)
}
\arguments{
\item{datExpr1}{is the control dat expression}

\item{datExpr2}{is the comparison expression, the network that we will test for conservation}

\item{modules}{is the color module vector in the control}
}
\value{
z scores
}
\description{
returns a Z score summary of the module preservation. Performs permuatation testing of the modules. This measures how well the preservation of grouping of modules of the control match in the comparison.
}
