% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantilWiseAnalysis.R
\name{quantileWiseAnalysis}
\alias{quantileWiseAnalysis}
\title{splits quartiles of repeat expression}
\usage{
quantileWiseAnalysis(tcga, tx.Biotype = "Alu", how = c("cpm", "tpm"))
}
\description{
this will collapse repeats and split by quartile, and plot the patient annotations across quartiles
}

