% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiTpm.R
\name{quasiTpm}
\alias{quasiTpm}
\title{quasi normalization of TPM by bundling}
\usage{
quasiTpm(kexp, ensg = ensgID, sample = sampleID)
}
\arguments{
\item{kexp}{a stage level kexp}

\item{ensg}{an EnsgID}

\item{sample}{a sample column}
}
\value{
normalized tpm
}
\description{
resulting quasi-normalization of TPM of a bundle numbers should now have variances which behave similarly to those of counts. this may be good enough to use the bundle wise TPM quanties for batch normalization and other shenanigans, then multiply through by the original proportions in each sample to retrieve the corrected per-transcript values to do unsupervised work.  this might also address aspects of single cell noise.
}

