% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDE_driver.R
\name{queryDE_driver}
\alias{queryDE_driver}
\title{queries a Differential Expression gene list for drivers in a given module biotype association}
\usage{
queryDE_driver(DE.list = NULL, moduleColor = "blue",
  module.biotype.driver = c("Alu", "Endogenous Retrovirus", "ERV3", "ERVK",
  "ERVL", "L1", "L2"),
  path = "~/Documents/Arkas-Paper-Data/AML-bonemarrow-LSCs/patient-plot-data/wgcna_data/repeat_short_list_biotypes/enrichment/RObjects/",
  p.cutoff = 0.05)
}
\value{
data print out to screen
}
\description{
given a Differential Expression gene list it is a great interest to query the DE list for each driver identity.
}

