% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDF.R
\name{queryDF}
\alias{queryDF}
\title{queries a data frame for ISGs,NK ligands and DDX helicases}
\usage{
queryDF(daFr = NULL, geneSymbol_col = NULL, NKLigs = c("CLEC2B",
  "KIR3DL3", "CLEC4F", "CLEC2A", "KLRG2", "CLEC4G", "ULBP3", "RAET1E", "RAET1L",
  "CLEC4C", "CLEC2L", "KLRD1", "CLEC10A", "CLEC3A", "CLEC3B", "CLEC19A",
  "CLEC1A", "MICA", "RAET1G"), ISG = c("IFNGR1", "TLR2", "IFNAR2", "IFITM3",
  "IFNL4", "IFNLR1", "TLR9", "IFNL1", "CCL16", "TRIL", "CCL4", "TLR5", "IFNL2",
  "IFNL3", "CCL17", "CCL22", "TLR7", "TLR10", "IFITM5", "IFNB1", "CCL25",
  "CCL15", "IFRD2", "ISG20"), ISG.pre = c("IL", "DDX", "IFN", "TLR", "IFIT",
  "CCL", "ISG", "IRF", "NFKB"), NK.pre = c("CLEC", "KIR", "KLR", "RAET",
  "MIC", "CLEX", "ULBP"))
}
\arguments{
\item{daFr}{should either be a driver data frame of arkas GWA df}

\item{geneSymbol_col}{is the column named Gene.symbol from arkas of hgnc_symbol from repeatToolKit drivers}
}
\value{
information only, no object
}
\description{
this is a first draft to query a data frame that has a column df gene names to return information about ISG, NK ligands or DDX helicases
}

