% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusage_annotation_prep.R
\name{qusage_annotation_prep}
\alias{qusage_annotation_prep}
\title{creates a kexp data frame which has gene symbol rownames}
\usage{
qusage_annotation_prep(kexp, species = c("Homo.sapiens", "Mus.musculus"))
}
\value{
gene name data frame
}
\description{
used mainly for qusage calls which returns a data frame of gene counts with rownames as gene HUGO names
}
