% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusage_multi.R
\name{qusage_multi}
\alias{qusage_multi}
\title{run qusage over multiple gene sets}
\usage{
qusage_multi(kexp, dbDir = "~/Documents/Arkas-Paper-Data/MSigDB/",
  db = "c2.cp.kegg.v5.1.symbols.gmt", comparison = "LSC",
  control = "pHSC", points.max = 2^12, table.number = 15)
}
\arguments{
\item{kexp}{a 2 stage level kexp}

\item{dbDir}{path to MSIGDB}

\item{db}{char name of db}

\item{comparison}{character of Numerator}

\item{control}{denominator}

\item{points.max}{number of values to distribute into tdist}

\item{table.number}{number to report}
}
\value{
qusage table
}
\description{
since the WGCNA enrichment algorithm is 'experimental' it can only provide preliminary analyses; additionally it returns warnings regarding pvalue calculations on correlations.  qusage can be run to verify using KEGG, GO, and immuneSignatures.  Note that WGCNA:Go enrichment inputs Entrez IDs which is gene specific, so in this function we collapse by gene ids, and run qusage
}

