% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusageAnalysis.R
\name{qusageAnalysis}
\alias{qusageAnalysis}
\title{runs qusage to test enrichment of ENSEMBL IDs}
\usage{
qusageAnalysis(annotatedKexp, byLevel = c("transcript", "gene"),
  reactomeSpecies = c("Mus musculus", "Homo sapiens"))
}
\description{
rus qusage to find enriched sets
}

