% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusageDbLite.R
\name{qusageDbLite}
\alias{qusageDbLite}
\title{Create a wgcnaDbLite object (usually called by subclass constructors)}
\usage{
qusageDbLite(x, path = ".", ...)
}
\arguments{
\item{x}{an annotation instance of wgcnaDbLite}

\item{path}{where it lives (default: ".")}

\item{...}{additional parameters}
}
\value{
a TxDbLite object
}
\description{
Create a wgcnaDbLite object (usually called by subclass constructors)
}

