% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRestrictions.R
\name{qusageMean}
\alias{qusageMean}
\title{qusage mean activity}
\usage{
qusageMean(kexp = NULL, read.cutoff = read.cutoff, comparison = NULL,
  controls = NULL, geneSet = geneSet, paired = FALSE,
  geneSetPath = geneSetPath, how = c("cpm", "tpm"))
}
\description{
calculates the mean pathway activity for a gene set
}

