% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusageRun.R
\name{qusageRun}
\alias{qusageRun}
\title{runs qusage}
\usage{
qusageRun(cnts_mt = NULL,
  gmt.path = "~/Documents/Arkas-Paper-Data/MSigDB/MsigDb_all/",
  MsigDB = c("c1.all.v5.1.symbols.gmt", "c2.all.v5.1.symbols.gmt",
  "c4.all.v5.1.symbols.gmt", "c5.all.v5.1.symbols.gmt",
  "c6.all.v5.1.symbols.gmt", "c7.all.v5.1.symbols.gmt",
  "h.all.v5.1.symbols.gmt"), comparison = "pHSC", control = "LSC",
  module = NULL, plotTable = FALSE, paired = TRUE)
}
\value{
enrichment data set
}
\description{
a general method for calling qusage
}
