% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusageTablesFromWGCNA.R
\name{qusageTablesFromWGCNA}
\alias{qusageDbLiteMetadata}
\alias{qusageTablesFromWGCNA}
\title{qusage enrichme modules returned from WGCNA.  the goal is to then select a MSigDB gmt and take the module.tx_biotype of interest and create a corresponding qusage enrichment database.}
\usage{
qusageTablesFromWGCNA(kexp, verbose = TRUE, dbname = "wgcnaDBLite.sqlite",
  version = "1.0.0", Module.color = "brown",
  MsigDB = c("c1.all.v5.1.symbols.gmt", "c2.all.v5.1.symbols.gmt",
  "c4.all.v5.1.symbols.gmt", "c5.all.v5.1.symbols.gmt",
  "c6.all.v5.1.symbols.gmt", "c7.all.v5.1.symbols.gmt",
  "h.all.v5.1.symbols.gmt"), how = c("cpm", "tpm"),
  species = c("Homo.sapiens", "Mus.musculus"), comparison1 = "pHSC",
  comparison2 = "Blast", controls = "LSC", paired = TRUE)

qusageDbLiteMetadata(kexp, packageName, species = NULL, Ensmbl, how = how)
}
\arguments{
\item{kexp}{the kexp should have pHSC, LSC and blast.}

\item{verbose}{boolean for std out}

\item{dbname}{this is the db name for wgcnaDbLite with correlation db}

\item{version}{character, for the metadata information}

\item{Module.color}{this is a color needed to grab the trait names}

\item{MsigDB}{which db to use}

\item{how}{either cpm or tpm}

\item{species}{either Homo.sapiesn or Mus.musculus}

\item{comparison1}{pHSC, or charactoer for 2 group copmarison}

\item{comparison2}{Blast, or character for 2nd pairwise. this is required, if you only have a two group then let comparison1=comparison2.}

\item{controls}{LSC, or character for control group}

\item{paired}{for patient paired data}

\item{packageName}{the name of the annotation package to be built}

\item{genomeVersion}{name of genome assembly for coordinates, e.g. "GRCh38"}

\item{sourceFile}{name of FASTA file(s) whence it was built, as a string}
}
\value{
a qusageDbLite db

a data.frame of metadata suitable for cramming into the database
}
\section{Functions}{
\itemize{
\item \code{qusageDbLiteMetadata}: create metadata for an qusageDbLite instance
}}

