% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameRepeatModuleColors.R
\name{renameRepeatModuleColors}
\alias{renameRepeatModuleColors}
\title{renames repeat module colors to the top three correlations with biotypes}
\usage{
renameRepeatModuleColors(rnames = NULL,
  rdbName = "wrcnaDbLite.cpm.6.sqlite",
  wgcnaDbName = "wgcnaDbLite.cpm.sqlite", geneModules = NULL, MEs = NULL)
}
\arguments{
\item{rnames}{the object returned from wrcna method}

\item{return}{returns rnames object with traitCorRenamed object}
}
\description{
the rnames object is assigned color names, and we want to rename the repeat module colors after the top 3 highest correlated tx_biotype groups to assign an identity.  this will systematically rename the repeat module colors to leading biotype. this script is primarily a helper function to moduleWiseAnalysis that aids in renaming gene modules to the leading correlated biotype families.
}

