% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatGeneBiotypeColorPalette.R
\name{repeatGeneBiotypeColorPalette}
\alias{repeatGeneBiotypeColorPalette}
\title{this will uniformly handle the color palette for repeat plotting}
\usage{
repeatGeneBiotypeColorPalette(kexp)
}
\arguments{
\item{kexp}{a kallisto experiment annotated}
}
\value{
a vector ready for rowAnnotation ComplexHeatmap
}
\description{
attempts to uniformly control the color assignment for biotype and family color plots for heatmapping
}

