% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatOdds.R
\name{repeatOdds}
\alias{repeatOdds}
\title{this will tabulate the enriched pathways as a proportion}
\usage{
repeatOdds(qusageDb = "qusageDbLite.cpm.sqlite",
  inflammatory = c("HALLMARK_IL6_JAK_STAT3_SIGNALING",
  "HALLMARK_INTERFERON_ALPHA_RESPONSE", "HALLMARK_INTERFERON_GAMMA_RESPONSE",
  "HALLMARK_INFLAMMATORY_RESPONSE", "HALLMARK_IL2_STAT5_SIGNALING"),
  tx.Biotypes = c("Alu", "DNA transposon", "Endogenous Retrovirus", "ERV1",
  "ERV3", "ERVK", "ERVL", "L1", "L2", "LTR Retrotransposon", "Satellite"),
  contrast = c("phsc", "blast"), paired = FALSE)
}
\arguments{
\item{qusageDb}{the dbname of the qusageDbLite}

\item{inflammatory}{the pathways related to MSigDb call}

\item{tx.Biotypes}{the cross correlated traits of interest}

\item{contrast}{phsc, or blast,  when tabulating phsc and blast data the test because paired,  when looking at individual contrasts we have independence}
}
\value{
a data frame of proportions between (inflammatory response,non inflammatory response) and proportions for each tx.Biotypes
}
\description{
the null hypothesis is that within each repeat biotype the proportions of inflammation response are identical to inflammation respsonse and non-inflammation response pathways the proportions are identical.  do all repeat biotypes associate equally to inflammation response?  this scripte will query the enrichment database , and tabulate the sum of inflammation related and non-inflammation releated enrichments.  note taht these enrichments were calculated as the most significantly cross-correlated genes for each partition module.  this is for an independent Chi-Square Test
}

