% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzekME.R, R/annotateDrivers.R, R/repeatTablesFromWGCNA.R
\name{getEntrezIDs}
\alias{getEntrezIDs}
\alias{getSymbols}
\alias{repeatTablesFromWGCNA}
\alias{wgcnaDbLiteMetadata}
\title{creates a WGCNA annotation package that includes many tables}
\usage{
getEntrezIDs(gxs, species)

getSymbols(gxs, species)

getEntrezIDs(gxs, species)

getSymbols(gxs, species)

repeatTablesFromWGCNA(kexp, lnames = NULL, useBiCor = TRUE,
  verbose = TRUE, dbname = "wgcnaDBLite", annotate = FALSE,
  version = "1.0.0", byWhich = c("gene", "repeat"))

getEntrezIDs(gxs, species)

getSymbols(gxs, species)

wgcnaDbLiteMetadata(kexp, packageName, species = NULL, Ensmbl, how = how,
  byWhich = byWhich)
}
\arguments{
\item{gxs}{geneIDs from a dataframe}

\item{species}{what kind of organism these genes are from}

\item{kexp}{any kexp doesn't matter, only use the metadata}

\item{lnames}{is the saved object from wgcna.R that includes basic formatted data compatible with WGCNA, datExpr, datTraits, annot library etc. Note that datTraits will be NA if considering gene modules only.}

\item{useBiCor}{boolean ,  if TRUE uses mid-correlation mid-weight robust to outliers}

\item{verbose}{boolean true for std out}

\item{dbname}{character}

\item{annotate}{boolean, for genes level set to TRUE, repeats set to FALSE}

\item{version}{character for meta}

\item{byWhich}{character}

\item{packageName}{the name of the annotation package to be built}

\item{gxs}{geneIDs from a dataframe}

\item{species}{what kind of organism these genes are from}

\item{gxs}{geneIDs from a dataframe}

\item{species}{what kind of organism these genes are from}

\item{genomeVersion}{name of genome assembly for coordinates, e.g. "GRCh38"}

\item{sourceFile}{name of FASTA file(s) whence it was built, as a string}
}
\value{
entrez_id values for the genes, where found

symbols for the genes, where found

entrez_id values for the genes, where found

symbols for the genes, where found

a SQLlite Data base

entrez_id values for the genes, where found

symbols for the genes, where found

a data.frame of metadata suitable for cramming into the database
}
\description{
after WGCNA creates a data object post networking and partitioning of data tables can add structure to the many gene~module and gene~trait correlations both including student-t and exact Fisher p.values with Gene annotations.  The tables created will be one table for each gene~module collection, and one table for each gene~trait collection with annotations. NOTE: wgcna.R needs to be called and the network object needs to be created prior the database tables.Further if you are not building a data base with phenotypic traits, then the datTraits will be NA.
}
\section{Functions}{
\itemize{
\item \code{getEntrezIDs}: add EntrezGene IDs for Ensembl genes

\item \code{getSymbols}: add symbols for Ensembl genes

\item \code{getEntrezIDs}: add EntrezGene IDs for Ensembl genes

\item \code{getSymbols}: add symbols for Ensembl genes

\item \code{getEntrezIDs}: add EntrezGene IDs for Ensembl genes

\item \code{getSymbols}: 

\item \code{wgcnaDbLiteMetadata}: create metadata for an wgcnaDbLite instance
}}

