% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvStageWiseAnalysis.R
\name{ruvStageWiseAnalysis}
\alias{ruvStageWiseAnalysis}
\title{runs RUV to normalize stage wise expression}
\usage{
ruvStageWiseAnalysis(stageKexp, p.value = 0.05, cutoff = 1)
}
\arguments{
\item{stageKexp}{kexp at the stages}

\item{p.value}{numeric significant cutoff}
}
\description{
plots the heatmap of top byMAD repeats. shows the composition of biotypes of the significant repeats, plots the delta beeswarm plots of each stage to show a global trend of the changes among states.
}

