% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showModuleMembers.R
\name{showModuleMembers}
\alias{showModuleMembers}
\title{shows the genes associated with a module set}
\usage{
showModuleMembers(lnames, biocolor = "blue")
}
\arguments{
\item{lnames}{the output from wgcna}

\item{biocolor}{a color module of interest}
}
\value{
a data frame with the eigenmodule color selected with the correlation score and pvalue and associated gene description.  downstream of this output could include filtering, and expresison plots
}
\description{
given a gene module color, we can examine the gene Ids , gene description along side an overall module enrichment activity result.  Module enrichment shows the overall enriched pathway analysis, and won't give specifics about genes.  this will display a specific module enrichment result and the module members with gene information.
}
