% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCorrect.R
\name{simCorrect}
\alias{simCorrect}
\title{corrects the probability differences in simulated data for exp dist}
\usage{
simCorrect(simDF = dum, delta.max = 0.05, delta.min = -0.05,
  lamb = fit.exp$estimate["rate"], x = DC0h_mm1, true.prob = true.prob)
}
\description{
read title
}

