% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleModuleAnalysis.R
\name{singleModuleAnalysis}
\alias{singleModuleAnalysis}
\title{this analyzes correlations of repeat and gene modules of fixed interest}
\usage{
singleModuleAnalysis(kexp, lnames, rnames,
  wgcnaDbName = "wgcna.blast.DbLite.sqlite",
  wrcnaDbName = "wrcna.blast.DbLite.cpm.sqlite",
  qusageDbName = "qusage.blast.DbLite.cpm.sqlite", geneModules = NULL,
  read.cutoff = 4, how = "tpm", enrichmentCaller = c("go", "qusage"),
  comparison1 = "phsc", comparison2 = "blast", openDevice = FALSE,
  p.value = 0.1, design = metadata(kexp)$design,
  hubPath = "~/Documents/Arkas-Paper-Data/AML-bonemarrow-LSCs/patient-plot-data/wgcna_data/repeat_short_list_biotypes/Buen_TCGA_Normal_Karytype_Analysis_ens84_rep2103/GO-analysis/BuenRostro_GO_analysis/Re-Analysis-Blast-LSC/",
  hubFile = "top_65_HubGenes.AML.csv")
}
\arguments{
\item{kexp}{this should be the full stage kexp, could even be a tcga kexp}

\item{lnames}{is the wgcna.R output and rnames is the wrcna.R call}

\item{wgcnaDbName}{this is the sql db of the wgcna}

\item{wrcnaDbName}{sql db of the wrcna repeat db}

\item{qusageDbName}{this is the sql db of a qusage db. not required if there is a missing qusage db one can call goEnrich for preliminary analyses}

\item{geneModules}{this should be single gene module of interest.  we can examine one gene module's association to all the repeat modules.  so geneModule should be a single color character}

\item{how}{either cpm or tpm, tpm recommended for heatmap visualizations}

\item{enrichmentCaller}{either go or qusage, for some cases go enrichment is an interest to study}

\item{comparison1}{this is the first comparison group the comparison key is in lower case letters 'mds' etc.}

\item{comparison2}{this is the second comparison group, if there is only 1 comparison group, then leave this as "blast"; the method will query the 'blast' contrast and come up with an empty set and discard it.}

\item{openDevice}{boolean if true then the images are opened on a device along with printing to pdf.  should be false for exploring all colors}

\item{p.value}{an interger significance alpha level}

\item{repeat.cutoff}{the cpm floor threshold if the cut off filters out elements that were not filtered in the wgcna method call, this will fail.  you must match the read.cutoff in the wgcna data base.}
}
\description{
after running wgcna and wrcna we are interested in finding associaitons between one fixed gene with all the repeat modules.  given a fixed module of interest we correlate and test for comparability between every repeat module.  If there is no contrast between the gene modules, i.e. qusage contrasts are not present, then we use the wgcna_goEnrich caller
}

