% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageWiseAnalysis.R
\name{stageWiseAnalysis}
\alias{stageWiseAnalysis}
\title{stage wise analysis of clonal expression using TMM normalization}
\usage{
stageWiseAnalysis(kexp, stage1 = "pHSC", stage2 = "LSC", stage3 = "Blast",
  read.cutoff = 1, byLevel = c("transcript", "tx_biotype", "gene_biotype"),
  p.value = 0.05, fold.cutoff = 0.4)
}
\arguments{
\item{kexp}{a kallisto expriment at the repeat level stage}

\item{stage1}{first stage}

\item{stage2}{second stage}

\item{stage3}{third stage}

\item{read.cutoff}{integer floor bundle count}

\item{byLevel}{a character of the level for DE analysis}

\item{p.value}{numeric sig leve}

\item{fold.cutoff}{floor for logFC}
}
\description{
stage level differential expression for global monotonicity analysis using TMM library normalization
}

