% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmmNorm.R
\name{tmmNorm}
\alias{tmmNorm}
\title{TMM Normalization KallistoExperiment}
\usage{
tmmNorm(kexp)
}
\arguments{
\item{kexp}{a kallistoExperiment}
}
\value{
a kallistoExperiment
}
\description{
input a KallistoExperiment and TMM the library size to normalize sequence depth and store the normalized values in the kexp container. The library normalized assays are added to the kallistoExperiment object which can be used in a patientPlot
}

