% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topHubGenes.R
\name{topHubGenes}
\alias{topHubGenes}
\title{returns the central hub genes for analysis within a kexp}
\usage{
topHubGenes(kexp, lnames = NULL, topNumber = 20)
}
\arguments{
\item{kexp}{kexp that is annotated}

\item{lnames}{wgcna.R output}

\item{topNumber}{connectivity cutoff}
}
\value{
a list of genes and their connectivity
}
\description{
performs a connectivity analysis and returns the top connected genes within an experiment.
}

