% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_adjacencyHeatmap.R
\name{wgcna_adjacencyHeatmap}
\alias{wgcna_adjacencyHeatmap}
\title{plots adjacency heatmap used by wgcna_analysis}
\usage{
wgcna_adjacencyHeatmap(MEs, datTraits, biotype = c("ERV1", "ERV2",
  "Endogenous Retrovirus", "ERV3", "ERVL", "L1", "L2", "LTR Retrotransposon"))
}
\arguments{
\item{MEs}{module eigen genes}

\item{datTraits}{data traits}

\item{biotype}{all the biotypes of interest}
}
\value{
images
}
\description{
for analyzing a network we want to plot all the interesting biotypes and look at their individiual correlations and relationship to Module EigeneGenes.
}
