% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_Cormap.R
\name{wgcna_Cormap}
\alias{wgcna_Cormap}
\title{produces a correlation map given wgcna network data}
\usage{
wgcna_Cormap(lnames, read.cutoff = 2, plotDot = FALSE, recalc = FALSE)
}
\arguments{
\item{read.cutoff}{integer for min cutoff}
}
\value{
images of eigengenes
}
\description{
after calling wgcna the datExpr and module colors are used to create a correlation map image.
}

