% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_filter.R
\name{wgcna_filter}
\alias{wgcna_filter}
\title{filters low mean counts}
\usage{
wgcna_filter(datExpr, annot)
}
\arguments{
\item{datExpr}{from the wgcna object lnames}

\item{annot}{from the gene annotation call}
}
\value{
a data frame with p adjusted q value genes
}
\description{
often it is needed to filter counts, this does that
}

