% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_FunctionMap.R
\name{wgcna_FunctionMap}
\alias{wgcna_FunctionMap}
\title{produces a correlation map given wgcna network datai of specific modules and assigns a function name to the module enrichment activity}
\usage{
wgcna_FunctionMap(lnames, rnames, read.cutoff = 2, recalc = FALSE,
  how = how, keyWord = NULL, dbName = NULL, qdbName = NULL)
}
\arguments{
\item{lnames}{this is the results from wgcna.R}

\item{read.cutoff}{integer for min cutoff}

\item{recalc}{boolean if truen then will recalculate the bicor and pvalues}

\item{plotDot}{boolean this plots the median correlation score for each biotypes in a line plot}

\item{targets}{this is a character vector of modules of interest to subset the corMap}
}
\value{
images of eigengenes
}
\description{
after calling wgcna the datExpr and module colors are used to create a correlation map image and a subset of these modules is plotted.
}

