% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_goEnrich.R
\name{wgcna_goEnrich}
\alias{wgcna_goEnrich}
\title{GO enrichment for wgcna_analysis modules}
\usage{
wgcna_goEnrich(datExpr, datTraits, annot, intModules = c("blue", "steelblue",
  "darkturquoise"))
}
\description{
downstream from wgcna_analysis we can investigate enrichment
}

