% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_plotAll_dendrograms.R
\name{wgcna_plotAll_dendrograms}
\alias{wgcna_plotAll_dendrograms}
\title{this plots all the dendrograms from a net or bwnet call}
\usage{
wgcna_plotAll_dendrograms(bwnet = NULL, whichWGCNA = c("single", "block"),
  bwModuleColors = NULL, bwLabels = NULL, how = "cpm", byWhich = "gene")
}
\arguments{
\item{whichWGCNA}{single or block level analysis, block is a bit less rigid}

\item{bwnet$dendrograms}{this is a gene tree from the block net call}

\item{net$dendrograms}{a gene tree from single net call}
}
\value{
cool network images and stuff
}
\description{
for wgcna.R call, the network dendrograms are clusterd based on average expression, one for each module. this iteratively goes through each and plots them.
}

