% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_scatterMod.R
\name{wgcna_scatterMod}
\alias{wgcna_scatterMod}
\title{plots a scatter correlation image of a fixed biotype and fixed eigengene module color}
\usage{
wgcna_scatterMod(lnames, biocolor = "blue", biotype = "ERVL")
}
\arguments{
\item{lnames}{a list of the results from the call wgcna method}

\item{biocolor}{the color of eigengene module}

\item{biotype}{the tx biotype of interest}
}
\value{
images of plot
}
\description{
allows for exploration of data presented in wgcna_Cormap
}

