% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_significance.R
\name{wgcna_significance}
\alias{wgcna_significance}
\title{this runs through significance analysis of modules}
\usage{
wgcna_significance(lnames, y = NULL, n.modules = 7)
}
\arguments{
\item{lnames}{the output from lnames wgcna call}
}
\value{
significance images
}
\description{
this will analyze the significance of modules returning a heatmap, and average gene significance per module
}
