% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_Tomplot.R
\name{wgcna_Tomplot}
\alias{wgcna_Tomplot}
\title{plots topological overlap matrices}
\usage{
wgcna_Tomplot(lnames, softPower = 10)
}
\arguments{
\item{lnames}{an object from wgcna method call}

\item{softPower}{integer 6 or 7 does nice}
}
\value{
images of TOM
}
\description{
plots the TOM plot given the wgcna matrices
}

