% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcnaConsensus.R
\name{wgcnaConsensus}
\alias{wgcnaConsensus}
\title{runs a consensus network analysis at the gene level to investigate repeat initiating pathways}
\usage{
wgcnaConsensus(kexp1, kexp2, read.cutoff = 2, minBranch = 2,
  whichWGCNA = c("single", "block"), entrezOnly = FALSE,
  species = c("Homo.sapiens", "Mus.musculus"), selectedPower = NULL,
  intBiotypes = c("Alu", "DNA transposon", "Endogenous Retrovirus", "ERV1",
  "ERV3", "ERVK", "ERVL", "L1", "L2", "LTR Retrotransposon", "Satellite"),
  useAllBiotypes = FALSE, tmm.norm = TRUE, useBiCor = TRUE,
  setLabels = c("kexp1", "kexp2"))
}
\arguments{
\item{kexp1}{a kexp one group stage is preferred}

\item{kexp2}{a kexp for consensus analysis}

\item{read.cutoff}{integer floor filter}

\item{minBranch}{integer for cluter min}

\item{whichWGCNA}{character single or block analysis, block is more sensitive}

\item{entrezOnly}{boolean, soon to be deprecated because entrez is auto filtered when enrichment testing}

\item{species}{char, mouse or humans}

\item{selectedPower}{6 usually is good. can rerun if NULL}

\item{intBiotypes}{character the tx_biotypes of interest}

\item{useAllBiotypes}{boolean if false then intBiotypes are used, if true than the correlations are checked against all tx_biotypes}
}
\value{
images and cluster at the gene and repeat level
}
\description{
runs WGNCA consensus analysis could find the gene correlation patterns to infer a relationship behind repeat activation. This uses the recommended 'biocor' function which is a bi-weight mid-correlation calculation. For normalization, the default uses tmm normalization and a log2 transformation for each cpm for genes, and repeat txBiotypes in the *same* way but on separate calls (this seems ok intuitively).  We use 'signed' networks based on the FAQ.  This will create the blockwise/single consensus module data frame and run the soft-thresholding and create a correlation heatmap.  the downstream method is wgcna_analsyis which investigates specific module color and specific biotype. Note this inputs 2 kallistoExperiments and creates a multiExpression structure as a necessary pre-processing step for blockWiseConsensus network call.
}
