% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcna.R
\name{wrcna}
\alias{wrcna}
\title{creates a network module at the repeat level to investigate repeat initiating pathways}
\usage{
wrcna(kexp, read.cutoff = 2, minBranch = 2, whichWGCNA = c("single",
  "block"), species = c("Homo.sapiens", "Mus.musculus"), selectedPower = 6,
  intBiotypes = c("acromeric", "centromeric", "CR1", "Alu", "DNA transposon",
  "Endogenous Retrovirus", "ERV1", "ERV3", "ERVK", "ERVL", "hAT", "HSFAU", "L1",
  "L2", "LTR Retrotransposon", "Eutr1", "Merlin", "PiggyBac", "Pseudogene",
  "Repetitive element", "satellite", "snRNA", "SVA", "TcMar", "telo",
  "Transposable Element", "Satellite"), useAllBiotypes = FALSE,
  tmm.norm = TRUE, useBiCor = TRUE, how = c("cpm", "tpm"),
  batchNormalize = FALSE, batchVector = NULL, design = NULL,
  copyNormalize = TRUE)
}
\arguments{
\item{kexp}{a kexp 2 group stage is preferred}

\item{read.cutoff}{integer floor filter}

\item{minBranch}{integer for cluter min}

\item{whichWGCNA}{character single or block analysis, block is more sensitive}

\item{species}{char, mouse or humans}

\item{selectedPower}{6 usually is good. can rerun if NULL}

\item{intBiotypes}{character the tx_biotypes of interest}

\item{useAllBiotypes}{boolean if false then intBiotypes are used, if true than the correlations are checked against all tx_biotypes}

\item{copyNormalize}{boolean, if true will execute copyNumberNormalize to normalize the repeat biotype counts by the copy number summation of the corresponding family type copy numbers find from repBase.}

\item{entrezOnly}{boolean, soon to be deprecated because entrez is auto filtered when enrichment testing}
}
\value{
images and cluster at the gene and repeat level
}
\description{
this differs from wgcna.R, where wgcna.R create phenotypic module analysis of genes and phenotypic repeat data.  wrna will run a single/block wise adjacency/correlation matrices of repeats into modules of transcripts and also returns an object ready for wgcnaDbLite that includes repeat phenotypic data. This uses the recommended 'biocor' function which is a bi-weight mid-correlation calculation. For normalization, the default uses tmm normalization and a log2 transformation for each cpm for genes,  We use 'signed' networks based on the FAQ.  This will create the blockwise/single module data frame and run the soft-thresholding and create a correlation heatmap.  the downstream method is wgcna_analsyis which investigates specific module color and specific biotype (phenotypic relationship).  tx_biotype is included as phenotypic data
}
