% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beeColumns.R
\name{beeColumns}
\alias{beeColumns}
\title{plots beeswarm of all repeats by clonal stage}
\usage{
beeColumns(kexp, patientID = NULL, repeat_biotype = NULL, selected = 20,
  what = c("tpm", "counts"))
}
\arguments{
\item{kexp}{a kexp}

\item{patientID}{a character to select patient info}

\item{repeat_biotype}{a character either 'element','LTR','SINE', etc}

\item{selected}{how many byMads to select}

\item{what}{is count or tpm}
}
\description{
plots beeswarm of all repeats by clonal stage
}

