% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHeatmap.R
\name{drawHeatmap}
\alias{drawHeatmap}
\title{Draws a heatmap with annotations}
\usage{
drawHeatmap(kexp, tags = NULL, annotations = TRUE, byType = c("counts",
  "tpm"), cutoff = 2)
}
\arguments{
\item{kexp}{a kallistoExperiment}

\item{tags}{character of top tags from DE list}

\item{annotations}{boolean  to include annotation side bar, or mutation type}

\item{cutoff}{integer  cutoff min}
}
\description{
draws a heatmap with annotations useful
}

