% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModel.R
\name{findModel}
\alias{findModel}
\title{find the repeats falling into orders}
\usage{
findModel(kexp, clones = NULL, hypothesis = NULL, alternative = NULL,
  patientID = NULL, repeatType = NULL)
}
\arguments{
\item{kexp}{a kallistoExperiment}

\item{clones}{preferable the column names of a kexp}

\item{hypothesis}{an ordering character}

\item{alternative}{an alternative ordering character}

\item{patientID}{a patient character identifier}

\item{repeatType}{a repeat character biotype}
}
\value{
a repeat listing for each hypothesis
}
\description{
find the repeats fallingo into clonal ordering with the alternative model and hypothesis model input for element selection.
}

