% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModel.R
\name{findModel}
\alias{findModel}
\title{find the repeats falling into orders}
\usage{
findModel(kexp, stage1 = "pHSC", stage2 = "LSC", stage3 = "Blast",
  patientID = NULL, repeatType = NULL, read.cutoff = 1,
  outputReport = FALSE, outputDir = NULL, what = c("counts", "tpm"))
}
\arguments{
\item{kexp}{a kallistoExperiment of a clonal patient kexp repeat}

\item{patientID}{a patient character identifier}

\item{repeatType}{a repeat character biotype}

\item{read.cutoff}{integer a min cutoff}

\item{outputReport}{boolean if true print a CSV}

\item{outputDir}{character path}

\item{hypothesis}{an ordering character}

\item{alternative}{an alternative ordering character}
}
\value{
a repeat listing for each hypothesis
}
\description{
find the repeats fallingo into clonal ordering with the alternative model and hypothesis model input for element selection. this has no p value association but expression pattern overview
}

