% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneTest.R
\name{geneTest}
\alias{geneTest}
\title{this creates an expression time plot of match patients}
\usage{
geneTest(kexp, geneName = "EVI", how = c("cpm", "tpm"), saveOut = FALSE,
  numberPairs = c(1, 2, 3), read.cutoff = 1)
}
\arguments{
\item{kexp}{this is a kexp for three stages patient wise, with 3 time-stages}
}
\value{
images and a pdf/jpeg
}
\description{
for paired time series, this plots the expression across time points in units of TPM. This plots each repeat biotype and every patient trio across stages.
}

