% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goof.R
\name{goof}
\alias{goof}
\title{goodness of fit for repeat data to poisson model or linear}
\usage{
goof(kexp, n = 100, numberOut = 999, n.trials = 10000, dataMat = NULL,
  bySamples = FALSE)
}
\arguments{
\item{kexp}{a kallisto experiment of repeat stage or all transcript stage}

\item{n}{the number of samples to draw from the poisson distribution}

\item{numberOut}{the number to plot out}

\item{n.trials}{the number of trials to create and sample from}

\item{bySamples}{boolean if true the good fit is calculate}

\item{datamat}{a matrix like object in the case where a kexp is undesired}
}
\value{
images to test poisson hyp
}
\description{
goodness of fit for repeat data follow simulating from a poisson and from normal distribution.  It first looks at all the samples together as one observation, then examines each sample individually to see how each sample fits.  the concern is that if all the samples are examined
}

