% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleHubGenes.R
\name{ModuleHubGenes}
\alias{ModuleHubGenes}
\title{returns the central hub genes for analysis within a kexp for a selected module color. topHubGenes finds hubs across all modules.}
\usage{
ModuleHubGenes(kexp, lnames = NULL, module.color = NULL,
  connectivity.cutoff = 0.6)
}
\arguments{
\item{kexp}{kexp that is annotated}

\item{lnames}{wgcna.R output}

\item{topNumber}{connectivity cutoff}
}
\value{
a list of genes and their connectivity
}
\description{
performs a connectivity analysis and returns the top connected genes within an experiment.
}
