% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusageDbLite.R
\name{qusageDbLite}
\alias{qusageDbLite}
\title{Constructor for an qusageDbLite object.}
\usage{
qusageDbLite(x, ...)
}
\arguments{
\item{x}{the name of the sqlite file}

\item{...}{any additional arguments relating to ENSEMBL annotation database.}
}
\value{
an qusageDbLite object.
}
\description{
Constructor for an qusageDbLite object.
}

