% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusageTablesFromWGCNA.R
\name{qusageTablesFromWGCNA}
\alias{qusageDbLiteMetadata}
\alias{qusageTablesFromWGCNA}
\title{qusage enrichme modules returned from WGCNA.  the goal is to then select a MSigDB gmt and take the module.tx_biotype of interest and create a corresponding qusage enrichment database.}
\usage{
qusageTablesFromWGCNA(kexp, useBiCor = TRUE, verbose = TRUE,
  dbname = "wgcnaDBLite.sqlite", annotate = TRUE, version = "1.0.0",
  Module.color = "brown", MsigDB = c("c1.all.v5.1.symbols.gmt",
  "c2.all.v5.1.symbols.gmt", "c4.all.v5.1.symbols.gmt",
  "c5.all.v5.1.symbols.gmt", "c6.all.v5.1.symbols.gmt",
  "c7.all.v5.1.symbols.gmt", "h.all.v5.1.symbols.gmt"), comparison = "pHSC",
  control = "LSC", how = c("cpm", "tpm"))

qusageDbLiteMetadata(kexp, packageName, species = NULL, Ensmbl, how = how)
}
\arguments{
\item{kexp}{the kexp should have pHSC, LSC and blast.}

\item{packageName}{the name of the annotation package to be built}

\item{genomeVersion}{name of genome assembly for coordinates, e.g. "GRCh38"}

\item{sourceFile}{name of FASTA file(s) whence it was built, as a string}
}
\value{
a data.frame of metadata suitable for cramming into the database
}
\section{Functions}{
\itemize{
\item \code{qusageDbLiteMetadata}: create metadata for an qusageDbLite instance
}}

