% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusageTablesFromWGCNA.R
\name{qusageTablesFromWGCNA}
\alias{qusageDbLiteMetadata}
\alias{qusageTablesFromWGCNA}
\title{qusage enrichme modules returned from WGCNA.  the goal is to then select a MSigDB gmt and take the module.tx_biotype of interest and create a corresponding qusage enrichment database.}
\usage{
qusageTablesFromWGCNA(kexp, verbose = TRUE, dbname = "wgcnaDBLite.sqlite",
  version = "1.0.0", Module.color = "brown",
  MsigDB = c("c1.all.v5.1.symbols.gmt", "c2.all.v5.1.symbols.gmt",
  "c4.all.v5.1.symbols.gmt", "c5.all.v5.1.symbols.gmt",
  "c6.all.v5.1.symbols.gmt", "c7.all.v5.1.symbols.gmt",
  "h.all.v5.1.symbols.gmt"), how = c("cpm", "tpm"),
  species = c("Homo.sapiens", "Mus.musculus"), comparison1 = "pHSC",
  comparison2 = "Blast", controls = "LSC", paired = TRUE,
  batchNormalize = FALSE, batchVector = NULL, comparisonNumber = 1,
  skipWithinEnrichment = TRUE, qusageDbName = "qusageDbLite")

qusageDbLiteMetadata(kexp, packageName, species = NULL, Ensmbl, how = how)
}
\arguments{
\item{kexp}{the kexp should have pHSC, LSC and blast.}

\item{verbose}{boolean for std out}

\item{dbname}{this is the db name for wgcnaDbLite with correlation db}

\item{version}{character, for the metadata information}

\item{Module.color}{this is a color needed to grab the trait names}

\item{MsigDB}{which db to use}

\item{how}{either cpm or tpm}

\item{species}{either Homo.sapiesn or Mus.musculus}

\item{comparison1}{pHSC, or charactoer for 2 group copmarison. these arguments are input into qusageRun and must be split using "_" with the factor of interest as the first leading term.}

\item{comparison2}{Blast, or character for 2nd pairwise. this is required, if you only have a two group then let comparison1=comparison2.  must be named using "_" with the factor of interest as the leading term.}

\item{controls}{LSC, or character for control group. these parameters are input into qusageRun with the factor of interest as the leading term.}

\item{paired}{for patient paired data}

\item{batchNormalize}{boolean, if true then the samples are from different batches and the edgeR batchCorrectEffect will be called for cpm}

\item{batchVector}{this is the batch vector. the length of the batch vector must match the number of columns. there are not any checks to validate the batch effect factors, user dependent input}

\item{comparisonNumber}{integer 1 or 2.  this integer reflects how many pairwise comparisons needed.  for time series analysis, it is sometimes desired to test for enrichment at N different time points,  intial-comparison1 , and intial-comparison2. qusage will be called to compare the comparisonNumber pairs}

\item{skipWithinEnrichment}{boolean, if true then only the gene modules are tested for pathway activation. if false then each biotype module iwth the highest significant genes correlating to the specific biotype module are additionally tested for pathway activation.}

\item{packageName}{the name of the annotation package to be built}

\item{genomeVersion}{name of genome assembly for coordinates, e.g. "GRCh38"}

\item{sourceFile}{name of FASTA file(s) whence it was built, as a string}
}
\value{
a qusageDbLite db

a data.frame of metadata suitable for cramming into the database
}
\section{Functions}{
\itemize{
\item \code{qusageDbLiteMetadata}: create metadata for an qusageDbLite instance
}}

