% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatTablesFromWGCNA.R
\name{repeatTablesFromWGCNA}
\alias{getEntrezIDs}
\alias{getSymbols}
\alias{repeatTablesFromWGCNA}
\alias{wgcnaDbLiteMetadata}
\title{creates a WGCNA annotation package that includes many tables}
\usage{
repeatTablesFromWGCNA(kexp, lnames, useBiCor = TRUE, verbose = TRUE,
  dbname = "wgcnaDBLite", annotate = TRUE, version = "1.0.0")

getEntrezIDs(gxs, species)

getSymbols(gxs, species)

wgcnaDbLiteMetadata(kexp, packageName, species = NULL, Ensmbl)
}
\arguments{
\item{lnames}{is the saved object from wgcna.R that includes basic formatted data compatible with WGCNA, datExpr, datTraits, annot library etc. FIX ME: have this annotation library be independent of wgcna.R}

\item{useBiCor}{boolean ,  if TRUE uses mid-correlation mid-weight robust to outliers}

\item{gxs}{geneIDs from a dataframe}

\item{species}{what kind of organism these genes are from}

\item{packageName}{the name of the annotation package to be built}

\item{genomeVersion}{name of genome assembly for coordinates, e.g. "GRCh38"}

\item{sourceFile}{name of FASTA file(s) whence it was built, as a string}
}
\value{
a SQLlite Data base

entrez_id values for the genes, where found

symbols for the genes, where found

a data.frame of metadata suitable for cramming into the database
}
\description{
after WGCNA creates a data object post networking and partitioning of data tables can add structure to the many gene~module and gene~trait correlations both including student-t and exact Fisher p.values with Gene annotations.  The tables created will be one table for each gene~module collection, and one table for each gene~trait collection with annotations.   FIX ME: there must be accessing methods for these tables.
}
\section{Functions}{
\itemize{
\item \code{getEntrezIDs}: add EntrezGene IDs for Ensembl genes

\item \code{getSymbols}: add symbols for Ensembl genes

\item \code{wgcnaDbLiteMetadata}: create metadata for an wgcnaDbLite instance
}}

