% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{wgcna_analysis}
\alias{wgcna_analysis}
\title{this analyzes downstream of wgcna}
\usage{
wgcna_analysis(lnames, biotype = c("ERV1", "ERV2", "Endogenous Retrovirus",
  "ERV3", "ERVL", "L1", "L2", "LTR Retrotransposon"), whichWGCNA = c("single",
  "block"), useBiCor = TRUE)
}
\arguments{
\item{lnames}{the output from the wgcna.R call}

\item{biotype}{names of datTraits}

\item{whichWGCNA}{single or block called}

\item{useBiCor}{boolean to use the mid-corrleation weighted robust to outliers}
}
\description{
this can investigate annotations of the datTraits object plots individual color modules and an individual biotype correlation and pvalue significance. the scatter plot returned is produced by the correlation value, p.value and regression line; it uses the base function lm to fit a linear model.
}

