% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{wgcna_analysis}
\alias{wgcna_analysis}
\title{this analyzes downstream of wgcna}
\usage{
wgcna_analysis(datExpr, datTraits, biotype = c("ERV1", "ERV2",
  "Endogenous Retrovirus", "ERV3", "ERVL", "L1", "L2", "LTR Retrotransposon"),
  biocolor = "blue", whichWGCNA = c("single", "block"))
}
\arguments{
\item{datExpr}{list entry from wgcna part 1 method [[1]]}

\item{datTraits}{list entry from wgcna part 1 method [[[2]]}

\item{biotype}{names of datTraits}
}
\description{
this can investigate annotations of the datTraits object plots individual color modules and an individual biotype correlation and pvalue significance
}

