% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_filter.R
\name{wgcna_filter}
\alias{wgcna_filter}
\title{filters low mean counts}
\usage{
wgcna_filter(lnames, geneTraitModuleDF, qvalCut = 0.1)
}
\arguments{
\item{lnames}{the object called from wgcna.R}

\item{geneTraitModuleDF}{the object called from wgcna_scatterMod.R}
}
\value{
a data frame with p adjusted q value genes
}
\description{
often it is needed to filter counts, this filters geneIDs from modules based on student-t p values.  we do have fisher pvalues, but not used to filter (yet)
}

