% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_Heatcor.R
\name{wgcna_Heatcor}
\alias{wgcna_Heatcor}
\title{produces a correlation map given wgcna network data using a different library ComplexHeatmap}
\usage{
wgcna_Heatcor(lnames = NULL, read.cutoff = 2, plotDot = FALSE,
  recalc = FALSE, targets = NULL, how = how, reOrder = TRUE)
}
\arguments{
\item{lnames}{this is the results from wgcna.R}

\item{read.cutoff}{integer for min cutoff}

\item{plotDot}{boolean this plots the median correlation score for each biotypes in a line plot}

\item{recalc}{boolean if truen then will recalculate the bicor and pvalues}

\item{targets}{this is a character vector of modules of interest to subset the corMap}

\item{orderBiotype}{either Alu or L1.}
}
\value{
images of eigengenes
}
\description{
after calling wgcna the datExpr and module colors are used to create a correlation map image with the columns ordered.
}

