% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcnaDbLite-class.R, R/wgcnaDbLite-methods.R
\docType{class}
\name{wgcnaDbLite-class}
\alias{dbconn,wgcnaDbLite-method}
\alias{metadata,wgcnaDbLite-method}
\alias{modulesBy,TxDbLite-method}
\alias{show,wgcnaDbLite-method}
\alias{wgcnaDbLite-class}
\title{A parent class for wgcna databases including annotations}
\usage{
\S4method{dbconn}{wgcnaDbLite}(x)

\S4method{show}{wgcnaDbLite}(object)

\S4method{metadata}{wgcnaDbLite}(x, ...)

\S4method{modulesBy}{TxDbLite}(x, Module.color = NULL,
  p.value.type = "studentT", trait = NULL)
}
\arguments{
\item{x}{wgcnaDblite SQL-lite database instance}

\item{object}{wgcnaDbLite SQL-lite annotation database}

\item{...}{additional parameters releated to annotation database}

\item{Module.color}{this is a selected table in the db, a module color}

\item{p.value.type}{either studentT or fisher, two types of p.value tests were done.}

\item{trait}{this is the trait of interest}

\item{x}{this is the SQLite db}

\item{x}{wgcna database}
}
\description{
A parent class for wgcna databases including annotations

A parent class for Ensembl based network annotations
}
\section{Slots}{

\describe{
\item{\code{con}}{a DBI connection}

\item{\code{tables}}{a list of annotation entries}
}}

